implment ImplicitMutCMap (formerly ImplicitCrdtMap)
