+++ client/src/crdt/map/index.ts
@@ -1,8 +1,9 @@
 export * from "./abstract_map";
 export * from "./decorated_map";
+export * from "./implicit_mut_map";
 export * from "./interfaces";
 export * from "./lww_map";
 export * from "./register_map";
 // register_mut_map is not exported since it's just internal
-export * from "./riak_crdt_maps";
+export * from "./resetting_mut_map";
 export * from "./tombstone_mut_map";
